테마 기능 통합 및 로컬 스토리지를 구현하라
* styled-components 대신 ThemeContext 를 도입하여 테마 상태 관리 개선했습니다.
* ThemeSwitcher 컴포넌트 추가하여 테마 전환 기능을 구현했습니다.
* 로컬 스토리지를 활용하여 테마 설정을 지속적으로 유지하도록 개선했습니다.
* App 컴포넌트에서 ThemeContext Provider를 사용하여 테마 정보를 전달합니다.
* 테마 변경 시 스타일 적용 및 로컬 스토리지 업데이트를 수행합니다.

변경된 파일:
* client/src/App.tsx
* client/src/components/header/ThemeSwitcher.tsx
* client/src/context/themeContext.tsx
