+++ client/src/App.tsx
@@ -1,24 +1,15 @@
 import Layout from './components/layout/Layout';
 import Home from './pages/Home';
-import { GlobalStyle } from './style/global';
-import { ThemeProvider } from 'styled-components';
-import { getTheme } from './style/theme';
+import { BookStoreThemeProvider } from './context/themeContext';
 import ThemeSwitcher from './components/header/ThemeSwitcher';
-import { useContext } from 'react';
-import { BookStoreThemeProvider, ThemeContext } from './context/themeContext';
 
 function App() {
-  const {themeName, setThemeName} = useContext(ThemeContext)
-  
   return (
     <BookStoreThemeProvider >
-      <ThemeProvider theme={getTheme(themeName)}>
-        <GlobalStyle themeName={themeName}/>
-        <ThemeSwitcher themeName={themeName} setThemeName={setThemeName}/>
+      <ThemeSwitcher />
         <Layout>
           <Home/>
         </Layout>
-      </ThemeProvider>
     </BookStoreThemeProvider>
   );
 }
