+++ client/src/context/themeContext.tsx
@@ -1,22 +1,43 @@
-import { ReactNode, createContext } from "react";
-import { ThemeName } from "../style/theme";
+import { ReactNode, createContext, useEffect, useState } from "react";
+import { ThemeName, getTheme } from "../style/theme";
+import { GlobalStyle } from "../style/global";
+import { ThemeProvider } from "styled-components";
 
+const DEFAULT_THEME_NAME = 'light'
+const THEME_LOCALSTORAGE_KEY = 'book_store_theme';
 interface State {
   themeName: ThemeName;
-  setThemeName: (themeName: ThemeName) => void,
+  toggleTheme: () => void,
 }
 
 export const state = {
-  themeName: 'light' as ThemeName,
-  setThemeName: (themeName: ThemeName) => {},
+  themeName: DEFAULT_THEME_NAME as ThemeName,
+  toggleTheme: () => {},
 };
 
 export const ThemeContext = createContext<State>(state)
 
 export const BookStoreThemeProvider = ({children}: {children: ReactNode}) => {
+  
+  const [themeName, setThemeName] = useState<ThemeName>('light');
+  
+  const toggleTheme = () => {
+    setThemeName(themeName === 'light' ? 'dark' : 'light')
+    localStorage.setItem(THEME_LOCALSTORAGE_KEY, 
+      themeName === 'light' ? 'dark' : 'light');
+  };
+
+  useEffect(() => {
+    const savedThemeName = localStorage.getItem(THEME_LOCALSTORAGE_KEY) as ThemeName;
+    setThemeName(savedThemeName || DEFAULT_THEME_NAME);
+  }, [])
+
   return (
-    <ThemeContext.Provider value={state} >
-      {children}
+    <ThemeContext.Provider value={{ themeName, toggleTheme }} >
+      <ThemeProvider theme={getTheme(themeName)}>
+        <GlobalStyle themeName={themeName}/>
+        {children}
+      </ThemeProvider>
     </ThemeContext.Provider>
   )
 }
